[NFC] Basics: add `AsyncBox` and `TokenBucket` helpers
